cleanup: removes shard reindexer v2
