+++ adapters/handlers/rest/configure_api.go
@@ -587,8 +587,6 @@ func MakeAppState(ctx context.Context, options *swag.CommandLineOptionsGroup) *s
 	reindexer := configureReindexer(appState, reindexCtx)
 	repo.WithReindexer(reindexer)
 
-	fmt.Printf("  ==> reindexer %+v\n\n", reindexer)
-
 	metaStoreReadyErr := fmt.Errorf("meta store ready")
 	metaStoreFailedErr := fmt.Errorf("meta store failed")
 	storeReadyCtx, storeReadyCancel := context.WithCancelCause(context.Background())
@@ -636,36 +634,6 @@ func MakeAppState(ctx context.Context, options *swag.CommandLineOptionsGroup) *s
 		reindexTaskNamesWithArgs["ShardInvertedReindexTask_SpecifiedIndex"] = appState.ServerConfig.Config.ReindexIndexesAtStartup
 	}
 
-	// reindexTasksV2Names := []string{}
-	// reindexTasksV2Args := map[string]any{}
-	// reindexFinishedV2 := make(chan error, 1)
-	// if appState.ServerConfig.Config.ReindexMapToBlockmaxAtStartup {
-	// 	reindexTasksV2Names = append(reindexTasksV2Names, "ShardInvertedReindexTask_MapToBlockmax")
-	// 	reindexTasksV2Args["ShardInvertedReindexTask_MapToBlockmax"] = map[string]any{
-	// 		"SwapBuckets":               appState.ServerConfig.Config.ReindexMapToBlockmaxConfig.SwapBuckets,
-	// 		"UnswapBuckets":             appState.ServerConfig.Config.ReindexMapToBlockmaxConfig.UnswapBuckets,
-	// 		"TidyBuckets":               appState.ServerConfig.Config.ReindexMapToBlockmaxConfig.TidyBuckets,
-	// 		"Rollback":                  appState.ServerConfig.Config.ReindexMapToBlockmaxConfig.Rollback,
-	// 		"ProcessingDurationSeconds": appState.ServerConfig.Config.ReindexMapToBlockmaxConfig.ProcessingDurationSeconds,
-	// 		"PauseDurationSeconds":      appState.ServerConfig.Config.ReindexMapToBlockmaxConfig.PauseDurationSeconds,
-	// 	}
-	// }
-
-	// if len(reindexTasksV2Names) > 0 {
-	// 	waitForMetaStore := func() error {
-	// 		// wait until meta store is ready, as reindex tasks need schema
-	// 		<-storeReadyCtx.Done()
-	// 		if err := context.Cause(storeReadyCtx); !errors.Is(err, metaStoreReadyErr) {
-	// 			return err
-	// 		}
-	// 		return nil
-	// 	}
-	// 	if err := runReindexerV2(reindexCtx, waitForMetaStore, migrator, appState.Logger, reindexTasksV2Names,
-	// 		reindexTasksV2Args, reindexFinishedV2); err != nil {
-	// 		os.Exit(1) // fail only in case of error (effectively when reindexer is not created)
-	// 	}
-	// }
-
 	if len(reindexTaskNamesWithArgs) > 0 {
 		// start reindexing inverted indexes (if requested by user) in the background
 		// allowing db to complete api configuration and start handling requests
@@ -704,7 +672,7 @@ func configureReindexer(appState *state.State, reindexCtx context.Context) db.Sh
 	logger := appState.Logger.WithField("action", "reindexV3")
 
 	if cfg := appState.ServerConfig.Config; cfg.ReindexMapToBlockmaxAtStartup {
-		tasks = append(tasks, db.NewShardInvertedReindexTaskMapToBlockmaxV3(
+		tasks = append(tasks, db.NewShardInvertedReindexTaskMapToBlockmax(
 			logger,
 			cfg.ReindexMapToBlockmaxConfig.SwapBuckets,
 			cfg.ReindexMapToBlockmaxConfig.UnswapBuckets,
@@ -726,85 +694,6 @@ func configureReindexer(appState *state.State, reindexCtx context.Context) db.Sh
 	reindexer.Init()
 	return reindexer
 }
-
-// func runReindexerV2(reindexCtx context.Context, waitForMetaStore func() error,
-// 	migrator *db.Migrator, logger logrus.FieldLogger,
-// 	reindexTasksV2Names []string, reindexTasksV2Args map[string]any, reindexFinishedV2 chan<- error,
-// ) error {
-// 	logger = logger.WithField("action", "reindexV2")
-
-// 	reindexer, err := migrator.ReindexerV2(reindexTasksV2Names, reindexTasksV2Args)
-// 	if err != nil {
-// 		logger.WithError(err).Error("creating reindexer")
-// 		return err
-// 	}
-
-// 	if reindexer.HasOnBefore() {
-// 		start := time.Now()
-// 		logger.Debug("starting waiting for meta store")
-// 		if err := waitForMetaStore(); err != nil {
-// 			err = fmt.Errorf("meta store not available: %w", err)
-// 			logger.WithField("took", time.Since(start)).WithError(err).Error("finished waiting for meta store")
-
-// 			reindexFinishedV2 <- err
-// 			return nil
-// 		}
-// 		logger.WithField("took", time.Since(start)).Debug("finished waiting for meta store")
-
-// 		start = time.Now()
-// 		logger.Info("starting on before")
-// 		if err := reindexer.OnBefore(reindexCtx); err != nil {
-// 			err = fmt.Errorf("on before: %w", err)
-// 			logger.WithField("took", time.Since(start)).WithError(err).Error("finished on before")
-// 		} else {
-// 			logger.WithField("took", time.Since(start)).Info("finished on before")
-// 		}
-
-// 		// continue even in onBefore failed on some tasks.
-// 		// remaining tasks will proceed with reindexing.
-// 		enterrors.GoWrapper(func() {
-// 			start = time.Now()
-// 			logger.Info("starting reindexing")
-// 			if err := reindexer.Reindex(reindexCtx); err != nil {
-// 				err = fmt.Errorf("reindex: %w", err)
-// 				logger.WithField("took", time.Since(start)).WithError(err).Error("finished reindexing")
-
-// 				reindexFinishedV2 <- err
-// 				return
-// 			}
-// 			logger.WithField("took", time.Since(start)).Info("finished reindexing")
-// 			reindexFinishedV2 <- nil
-// 		}, logger)
-
-// 		return nil
-// 	}
-
-// 	enterrors.GoWrapper(func() {
-// 		start := time.Now()
-// 		logger.Debug("starting waiting for meta store")
-// 		if err := waitForMetaStore(); err != nil {
-// 			err = fmt.Errorf("meta store not available: %w", err)
-// 			logger.WithField("took", time.Since(start)).WithError(err).Error("finished waiting for meta store")
-
-// 			reindexFinishedV2 <- err
-// 			return
-// 		}
-// 		logger.WithField("took", time.Since(start)).Debug("finished waiting for meta store")
-
-// 		start = time.Now()
-// 		logger.Info("starting reindexing")
-// 		if err := reindexer.Reindex(reindexCtx); err != nil {
-// 			err = fmt.Errorf("reindex: %w", err)
-// 			logger.WithField("took", time.Since(start)).WithError(err).Error("finished reindexing")
-
-// 			reindexFinishedV2 <- err
-// 			return
-// 		}
-// 		logger.WithField("took", time.Since(start)).Info("finished reindexing")
-// 	}, logger)
-
-// 	return nil
-// }
 
 func parseNode2Port(appState *state.State) (m map[string]int, err error) {
 	m = make(map[string]int, len(appState.ServerConfig.Config.Raft.Join))
