+++ adapters/repos/db/inverted_reindexer_map_to_blockmax.go
@@ -22,43 +22,27 @@ import (
 	"sync"
 	"time"
 
+	"github.com/google/uuid"
+	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"github.com/weaviate/weaviate/adapters/repos/db/helpers"
 	"github.com/weaviate/weaviate/adapters/repos/db/inverted"
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv"
+	"github.com/weaviate/weaviate/entities/additional"
 	"github.com/weaviate/weaviate/entities/concurrency"
 	enterrors "github.com/weaviate/weaviate/entities/errors"
 	"github.com/weaviate/weaviate/entities/storobj"
 )
 
-type mapToBlockmaxConfig struct {
-	swapBuckets                   bool
-	unswapBuckets                 bool
-	tidyBuckets                   bool
-	rollback                      bool
-	concurrency                   int
-	memtableOptBlockmaxFactor     int
-	processingDuration            time.Duration
-	pauseDuration                 time.Duration
-	checkProcessingEveryNoObjects int
-}
-
-type ShardInvertedReindexTask_MapToBlockmax struct {
-	logger               logrus.FieldLogger
-	newReindexTracker    func(lsmPath string) (mapToBlockmaxReindexTracker, error)
-	keyParser            indexKeyParser
-	objectsIteratorAsync objectsIteratorAsync
-	config               mapToBlockmaxConfig
-}
-
 func NewShardInvertedReindexTaskMapToBlockmax(logger logrus.FieldLogger,
 	swapBuckets, unswapBuckets, tidyBuckets, rollback bool,
 	processingDuration, pauseDuration time.Duration,
-) *ShardInvertedReindexTask_MapToBlockmax {
+) *ShardReindexTask_MapToBlockmax {
+	name := "MapToBlockmax"
 	keyParser := &uuidKeyParser{}
 	objectsIteratorAsync := uuidObjectsIteratorAsync
 
-	logger = logger.WithField("task", "MapToBlockmax")
+	logger = logger.WithField("task", name)
 	newReindexTracker := func(lsmPath string) (mapToBlockmaxReindexTracker, error) {
 		rt := newFileMapToBlockmaxReindexTracker(lsmPath, keyParser)
 		if err := rt.init(); err != nil {
@@ -67,7 +51,7 @@ func NewShardInvertedReindexTaskMapToBlockmax(logger logrus.FieldLogger,
 		return rt, nil
 	}
 
-	return &ShardInvertedReindexTask_MapToBlockmax{
+	return &ShardReindexTask_MapToBlockmax{
 		logger:               logger,
 		newReindexTracker:    newReindexTracker,
 		keyParser:            keyParser,
@@ -86,39 +70,45 @@ func NewShardInvertedReindexTaskMapToBlockmax(logger logrus.FieldLogger,
 	}
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) HasOnBefore() bool {
-	return true
+type ShardReindexTask_MapToBlockmax struct {
+	name                 string
+	logger               logrus.FieldLogger
+	newReindexTracker    func(lsmPath string) (mapToBlockmaxReindexTracker, error)
+	keyParser            indexKeyParser
+	objectsIteratorAsync objectsIteratorAsync
+	config               mapToBlockmaxConfig
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) OnBefore(ctx context.Context) (err error) {
-	logger := t.logger.WithField("method", "OnBefore")
-	logger.Debug("starting")
-	defer func(started time.Time) {
-		logger = logger.WithField("took", time.Since(started))
-		if err != nil {
-			logger.WithError(err).Error("finished with error")
-		} else {
-			logger.Debug("finished")
-		}
-	}(time.Now())
-
-	return nil
+type mapToBlockmaxConfig struct {
+	swapBuckets                   bool
+	unswapBuckets                 bool
+	tidyBuckets                   bool
+	rollback                      bool
+	concurrency                   int
+	memtableOptBlockmaxFactor     int
+	processingDuration            time.Duration
+	pauseDuration                 time.Duration
+	checkProcessingEveryNoObjects int
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) OnBeforeByShard(ctx context.Context, shard ShardLike) (err error) {
+func (t *ShardReindexTask_MapToBlockmax) Name() string {
+	return t.name
+}
+
+func (t *ShardReindexTask_MapToBlockmax) OnBeforeLsmInit(ctx context.Context, shard *Shard) (err error) {
 	collectionName := shard.Index().Config.ClassName.String()
 	logger := t.logger.WithFields(map[string]any{
 		"collection": collectionName,
 		"shard":      shard.Name(),
-		"method":     "OnBeforeShard",
+		"method":     "OnBeforeLsmInit",
 	})
-	logger.Debug("starting")
+	logger.Info("starting")
 	defer func(started time.Time) {
 		logger = logger.WithField("took", time.Since(started))
 		if err != nil {
 			logger.WithError(err).Error("finished with error")
 		} else {
-			logger.Debug("finished")
+			logger.Info("finished")
 		}
 	}(time.Now())
 
@@ -128,12 +118,11 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) OnBeforeByShard(ctx context.Con
 		return
 	}
 
-	props, err := t.getPropsToReindex(shard, rt)
+	props, err := t.readPropsToReindex(rt)
 	if err != nil {
-		err = fmt.Errorf("getting reindexable props: %w", err)
+		err = fmt.Errorf("reading reindexable props: %w", err)
 		return
 	}
-	logger.WithField("props", props).Debug("props found")
 
 	if t.config.rollback {
 		logger.Debug("rollback started")
@@ -164,6 +153,7 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) OnBeforeByShard(ctx context.Con
 	}
 
 	if len(props) == 0 {
+		logger.Debug("no props read. nothing to do")
 		return nil
 	}
 
@@ -173,16 +163,14 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) OnBeforeByShard(ctx context.Con
 	}
 
 	isMerged := rt.isMerged()
-	if !isMerged {
-		if rt.isReindexed() {
-			logger.Debug("reindexed, not merged. merging buckets")
+	if !isMerged && rt.isReindexed() {
+		logger.Debug("reindexed, not merged. merging buckets")
 
-			if err = t.mergeReindexAndIngestBuckets(ctx, logger, shard, rt, props); err != nil {
-				err = fmt.Errorf("merging reindex and ingest buckets:%w", err)
-				return
-			}
-			isMerged = true
+		if err = t.mergeReindexAndIngestBuckets(ctx, logger, shard, rt, props); err != nil {
+			err = fmt.Errorf("merging reindex and ingest buckets:%w", err)
+			return
 		}
+		isMerged = true
 	}
 
 	if err = ctx.Err(); err != nil {
@@ -192,21 +180,20 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) OnBeforeByShard(ctx context.Con
 
 	isSwapped := rt.isSwapped()
 	isTidied := rt.isTidied()
-
-	if !isSwapped {
-		shard.markSearchableBlockmaxProperties(props...)
-	}
-
 	if isMerged {
 		if isSwapped {
-			if t.config.unswapBuckets && !isTidied {
-				logger.Debug("swapped, not tidied. unswapping buckets")
+			if t.config.unswapBuckets {
+				if isTidied {
+					logger.Debug("swapped and tidied. can not be unswapped")
+				} else {
+					logger.Debug("swapped, not tidied. unswapping buckets")
 
-				if err = t.unswapIngestAndMapBuckets(ctx, logger, shard, rt, props); err != nil {
-					err = fmt.Errorf("unswapping ingest and map buckets:%w", err)
-					return
+					if err = t.unswapIngestAndMapBuckets(ctx, logger, shard, rt, props); err != nil {
+						err = fmt.Errorf("unswapping ingest and map buckets:%w", err)
+						return
+					}
+					isSwapped = false
 				}
-				isSwapped = false
 			}
 		} else {
 			if t.config.swapBuckets {
@@ -227,39 +214,93 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) OnBeforeByShard(ctx context.Con
 	}
 
 	if isSwapped {
-		if !isTidied {
-			if t.config.tidyBuckets {
-				logger.Debug("swapped, not tidied. tidying buckets")
-
-				if err = t.tidyMapBuckets(ctx, logger, shard, rt, props); err != nil {
-					err = fmt.Errorf("tidying map buckets:%w", err)
-					return
-				}
-			} else {
-				logger.Debug("swapped, not tidied. starting map buckets")
-
-				if err = t.loadMapSearchBuckets(ctx, logger, shard, props); err != nil {
-					err = fmt.Errorf("starting map buckets:%w", err)
-					return
-				}
-				if err = t.duplicateToMapBuckets(ctx, logger, shard, props); err != nil {
-					err = fmt.Errorf("duplicating map buckets:%w", err)
-					return
-				}
-			}
-		} else {
+		if isTidied {
 			logger.Debug("tidied. nothing to do")
+			return nil
+		}
+
+		if t.config.tidyBuckets {
+			logger.Debug("swapped, not tidied. tidying buckets")
+
+			if err = t.tidyMapBuckets(ctx, logger, shard, rt, props); err != nil {
+				err = fmt.Errorf("tidying map buckets:%w", err)
+				return
+			}
+		}
+	}
+
+	return nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) OnAfterLsmInit(ctx context.Context, shard *Shard) (err error) {
+	collectionName := shard.Index().Config.ClassName.String()
+	logger := t.logger.WithFields(map[string]any{
+		"collection": collectionName,
+		"shard":      shard.Name(),
+		"method":     "OnAfterLsmInit",
+	})
+	logger.Info("starting")
+	defer func(started time.Time) {
+		logger = logger.WithField("took", time.Since(started))
+		if err != nil {
+			logger.WithError(err).Error("finished with error")
+		} else {
+			logger.Info("finished")
+		}
+	}(time.Now())
+
+	if t.config.rollback {
+		logger.Debug("rollback. nothing to do")
+		return nil
+	}
+
+	rt, err := t.newReindexTracker(shard.pathLSM())
+	if err != nil {
+		err = fmt.Errorf("creating reindex tracker: %w", err)
+		return
+	}
+
+	props, err := t.getPropsToReindex(shard, rt)
+	if err != nil {
+		err = fmt.Errorf("getting reindexable props: %w", err)
+		return
+	}
+	logger.WithField("props", props).Debug("props found")
+	if len(props) == 0 {
+		logger.Debug("no props found. nothing to do")
+		return nil
+	}
+
+	if rt.isSwapped() {
+		if !rt.isTidied() {
+			logger.Debug("swapped, not tidied. starting map buckets")
+
+			if err = t.loadMapSearchBuckets(ctx, logger, shard, props); err != nil {
+				err = fmt.Errorf("starting map buckets:%w", err)
+				return
+			}
+			if err = t.duplicateToMapBuckets(shard, props); err != nil {
+				err = fmt.Errorf("duplicating map buckets:%w", err)
+				return
+			}
 		}
 	} else {
-		logger.Debug("merged, not swapped. starting ingest buckets")
+		isMerged := rt.isMerged()
+		if isMerged {
+			logger.Debug("merged, not swapped. starting ingest buckets")
+		} else {
+			logger.Debug("not merged. starting ingest buckets")
+		}
+
+		shard.markSearchableBlockmaxProperties(props...)
 
 		// since reindex bucket will be merged into ingest bucket with reindex segments being before ingest,
-		// ingest segments should not be compacted and tombstones kept
+		// ingest segments should not be compacted and tombstones should be kept
 		if err = t.loadIngestSearchBuckets(ctx, logger, shard, props, !isMerged, !isMerged); err != nil {
 			err = fmt.Errorf("starting ingest buckets:%w", err)
 			return
 		}
-		if err = t.duplicateToIngestBuckets(ctx, logger, shard, props); err != nil {
+		if err = t.duplicateToIngestBuckets(shard, props); err != nil {
 			err = fmt.Errorf("duplicating ingest buckets:%w", err)
 			return
 		}
@@ -268,565 +309,28 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) OnBeforeByShard(ctx context.Con
 	return nil
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) mergeReindexAndIngestBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, rt mapToBlockmaxReindexTracker, props []string,
-) error {
-	lsmPath := shard.pathLSM()
-	segmentPathsToMove := [][2]string{}
-	bucketPathsToRemove := make([]string, 0, len(props))
-	lock := new(sync.Mutex)
-
-	eg, gctx := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range props {
-		propName := props[i]
-
-		eg.Go(func() error {
-			reindexBucketName := t.reindexBucketName(propName)
-			reindexBucketPath := filepath.Join(lsmPath, reindexBucketName)
-			ingestBucketName := t.ingestBucketName(propName)
-			ingestBucketPath := filepath.Join(lsmPath, ingestBucketName)
-
-			for {
-				propSegmentPathsToMove, needsRecover, err := t.getSegmentPathsToMove(reindexBucketPath, ingestBucketPath)
-				if err != nil {
-					return fmt.Errorf("buckets %q & %q: %w", reindexBucketName, ingestBucketName, err)
-				}
-
-				if needsRecover {
-					if err := t.recoverReindexBucket(gctx, logger, shard, reindexBucketName); err != nil {
-						return err
-					}
-				} else {
-					lock.Lock()
-					bucketPathsToRemove = append(bucketPathsToRemove, reindexBucketPath)
-					segmentPathsToMove = append(segmentPathsToMove, propSegmentPathsToMove...)
-					lock.Unlock()
-					return nil
-				}
-			}
-		})
-	}
-	if err := eg.Wait(); err != nil {
-		return err
-	}
-
-	logger.WithField("segments_to_move", segmentPathsToMove).WithField("buckets_to_remove", bucketPathsToRemove).
-		Debug("merging reindex and ingest buckets")
-
-	eg, gctx = enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range segmentPathsToMove {
-		i := i
-		eg.Go(func() error {
-			return os.Rename(segmentPathsToMove[i][0], segmentPathsToMove[i][1])
-		})
-	}
-	if err := eg.Wait(); err != nil {
-		return fmt.Errorf("moving segment: %w", err)
-	}
-
-	eg, gctx = enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range bucketPathsToRemove {
-		i := i
-		eg.Go(func() error {
-			return os.RemoveAll(bucketPathsToRemove[i])
-		})
-	}
-	if err := eg.Wait(); err != nil {
-		return fmt.Errorf("removing bucket: %w", err)
-	}
-
-	if err := rt.markMerged(); err != nil {
-		return fmt.Errorf("marking reindex merged: %w", err)
-	}
-	return nil
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) getSegmentPathsToMove(bucketPathSrc, bucketPathDst string,
-) ([][2]string, bool, error) {
-	segmentPaths := [][2]string{}
-	needsRecover := false
-
-	err := filepath.WalkDir(bucketPathSrc, func(path string, d os.DirEntry, err error) error {
-		if d.IsDir() {
-			return nil
-		}
-		if t.isSegmentDb(d.Name()) || t.isSegmentBloom(d.Name()) {
-			ext := filepath.Ext(d.Name())
-			id := strings.TrimSuffix(strings.TrimPrefix(d.Name(), "segment-"), ext)
-			timestamp, err := strconv.ParseInt(id, 10, 64)
-			if err != nil {
-				return err
-			}
-			timestampPast := time.Unix(0, timestamp).AddDate(-23, 0, 0).UnixNano()
-
-			segmentPaths = append(segmentPaths, [2]string{
-				path, filepath.Join(bucketPathDst, fmt.Sprintf("segment-%d%s", timestampPast, ext)),
-			})
-		} else if t.isSegmentWal(d.Name()) {
-			if info, err := d.Info(); err != nil {
-				return err
-			} else if info.Size() > 0 {
-				needsRecover = true
-				return filepath.SkipAll
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, false, err
-	}
-	if needsRecover {
-		return nil, true, nil
-	}
-	return segmentPaths, false, nil
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) swapIngestAndMapBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, rt mapToBlockmaxReindexTracker, props []string,
-) error {
-	lsmPath := shard.pathLSM()
-	store := shard.Store()
-	propsToSwap := make([]string, 0, len(props))
-
-	eg, gctx := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range props {
-		propName := props[i]
-
-		if !rt.isSwappedProp(props[i]) {
-			propsToSwap = append(propsToSwap, propName)
-
-			eg.Go(func() error {
-				bucketName := helpers.BucketSearchableFromPropNameLSM(propName)
-
-				logger.WithField("bucket", bucketName).Debug("shutting down bucket")
-				if err := store.ShutdownBucket(gctx, bucketName); err != nil {
-					return err
-				}
-				return nil
-			})
-		}
-	}
-	if err := eg.Wait(); err != nil {
-		return err
-	}
-
-	logger.Debug("shut down searchable buckets for swap")
-
-	// TODO aliszka:blockmax handle partially swapped buckets?
-	eg, gctx = enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range propsToSwap {
-		propName := propsToSwap[i]
-
-		eg.Go(func() error {
-			bucketName := helpers.BucketSearchableFromPropNameLSM(propName)
-			bucketPath := filepath.Join(lsmPath, bucketName)
-			ingestBucketName := t.ingestBucketName(propName)
-			ingestBucketPath := filepath.Join(lsmPath, ingestBucketName)
-			mapBucketName := t.mapBucketName(propName)
-			mapBucketPath := filepath.Join(lsmPath, mapBucketName)
-
-			logger.WithFields(map[string]any{
-				"bucket":        bucketName,
-				"ingest_bucket": ingestBucketName,
-				"map_bucket":    mapBucketName,
-			}).Debug("swapping buckets")
-
-			if err := os.Rename(bucketPath, mapBucketPath); err != nil {
-				return err
-			}
-			if err := os.Rename(ingestBucketPath, bucketPath); err != nil {
-				return err
-			}
-			if err := rt.markSwappedProp(propName); err != nil {
-				return fmt.Errorf("marking reindex swapped for %q: %w", propName, err)
-			}
-			return nil
-		})
-	}
-
-	if err := eg.Wait(); err != nil {
-		return err
-	}
-	if err := rt.markSwapped(); err != nil {
-		return fmt.Errorf("marking reindex swapped: %w", err)
-	}
-
-	logger.Debug("swapped searchable buckets")
-
-	if err := t.loadSwappedSearchBuckets(ctx, logger, shard, props); err != nil {
-		return err
-	}
-	logger.Debug("loaded searchable buckets after swap")
-
-	return nil
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) unswapIngestAndMapBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, rt mapToBlockmaxReindexTracker, props []string,
-) error {
-	lsmPath := shard.pathLSM()
-	store := shard.Store()
-	propsToUnswap := make([]string, 0, len(props))
-
-	eg, gctx := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range props {
-		propName := props[i]
-
-		if rt.isSwappedProp(props[i]) {
-			propsToUnswap = append(propsToUnswap, propName)
-
-			eg.Go(func() error {
-				bucketName := helpers.BucketSearchableFromPropNameLSM(propName)
-
-				logger.WithField("bucket", bucketName).Debug("shutting down bucket")
-				if err := store.ShutdownBucket(gctx, bucketName); err != nil {
-					return err
-				}
-				return nil
-			})
-		}
-	}
-	if err := eg.Wait(); err != nil {
-		return err
-	}
-
-	logger.Debug("shut down searchable buckets for unswap")
-
-	eg, gctx = enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range propsToUnswap {
-		propName := propsToUnswap[i]
-
-		eg.Go(func() error {
-			bucketName := helpers.BucketSearchableFromPropNameLSM(propName)
-			bucketPath := filepath.Join(lsmPath, bucketName)
-			ingestBucketName := t.ingestBucketName(propName)
-			ingestBucketPath := filepath.Join(lsmPath, ingestBucketName)
-			mapBucketName := t.mapBucketName(propName)
-			mapBucketPath := filepath.Join(lsmPath, mapBucketName)
-
-			logger.WithFields(map[string]any{
-				"bucket":        bucketName,
-				"ingest_bucket": ingestBucketName,
-				"map_bucket":    mapBucketName,
-			}).Debug("unswapping buckets")
-
-			if err := os.Rename(bucketPath, ingestBucketPath); err != nil {
-				return err
-			}
-			if err := os.Rename(mapBucketPath, bucketPath); err != nil {
-				return err
-			}
-			if err := rt.unmarkSwappedProp(propName); err != nil {
-				return fmt.Errorf("unmarking reindex swapped for %q: %w", propName, err)
-			}
-			return nil
-		})
-	}
-
-	if err := eg.Wait(); err != nil {
-		return err
-	}
-	if err := rt.unmarkSwapped(); err != nil {
-		return fmt.Errorf("unmarking reindex swapped: %w", err)
-	}
-
-	logger.Debug("unswapped searchable buckets")
-
-	if err := t.loadUnswappedSearchBuckets(ctx, logger, shard, props); err != nil {
-		return err
-	}
-	logger.Debug("loaded searchable buckets after unswap")
-
-	return nil
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) tidyMapBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, rt mapToBlockmaxReindexTracker, props []string,
-) error {
-	lsmPath := shard.pathLSM()
-
-	eg, _ := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range props {
-		propName := props[i]
-
-		eg.Go(func() error {
-			bucketName := t.mapBucketName(propName)
-			bucketPath := filepath.Join(lsmPath, bucketName)
-			if err := os.RemoveAll(bucketPath); err != nil {
-				return err
-			}
-			return nil
-		})
-	}
-
-	if err := eg.Wait(); err != nil {
-		return err
-	}
-	if err := rt.markTidied(); err != nil {
-		return fmt.Errorf("marking reindex tidied: %w", err)
-	}
-	return nil
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) loadReindexSearchBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string,
-) error {
-	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyInverted, false, false, t.config.memtableOptBlockmaxFactor)
-	return t.loadBuckets(ctx, logger, shard, props, t.reindexBucketName, bucketOpts)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) loadIngestSearchBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string,
-	disableCompaction, keepTombstones bool,
-) error {
-	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyInverted, disableCompaction, keepTombstones, t.config.memtableOptBlockmaxFactor)
-	return t.loadBuckets(ctx, logger, shard, props, t.ingestBucketName, bucketOpts)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) loadMapSearchBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string,
-) error {
-	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyMapCollection, false, false, 1)
-	return t.loadBuckets(ctx, logger, shard, props, t.mapBucketName, bucketOpts)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) loadSwappedSearchBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string,
-) error {
-	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyInverted, false, false, 1)
-	return t.loadBuckets(ctx, logger, shard, props, helpers.BucketSearchableFromPropNameLSM, bucketOpts)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) loadUnswappedSearchBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string,
-) error {
-	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyMapCollection, false, false, 1)
-	return t.loadBuckets(ctx, logger, shard, props, helpers.BucketSearchableFromPropNameLSM, bucketOpts)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) loadBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string, bucketNamer func(string) string,
-	bucketOpts []lsmkv.BucketOption,
-) error {
-	store := shard.Store()
-
-	eg, gctx := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range props {
-		propName := props[i]
-
-		eg.Go(func() error {
-			bucketName := bucketNamer(propName)
-			logger.WithField("bucket", bucketName).Debug("loading bucket")
-			if err := store.CreateOrLoadBucket(gctx, bucketName, bucketOpts...); err != nil {
-				return err
-			}
-			logger.WithField("bucket", bucketName).Debug("bucket loaded")
-			return nil
-		})
-	}
-
-	if err := eg.Wait(); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) recoverReindexBucket(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, bucketName string,
-) error {
-	store := shard.Store()
-	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyInverted, true, false, t.config.memtableOptBlockmaxFactor)
-
-	logger.WithField("bucket", bucketName).Debug("loading bucket")
-	if err := store.CreateOrLoadBucket(ctx, bucketName, bucketOpts...); err != nil {
-		return fmt.Errorf("bucket %q: %w", bucketName, err)
-	}
-	logger.WithField("bucket", bucketName).Debug("shutting down bucket")
-	if err := store.ShutdownBucket(ctx, bucketName); err != nil {
-		return fmt.Errorf("bucket %q: %w", bucketName, err)
-	}
-	logger.WithField("bucket", bucketName).Debug("shut down bucket")
-
-	return nil
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) removeReindexBucketsDirs(ctx context.Context, logger logrus.FieldLogger,
-	shard ShardLike, props []string,
-) error {
-	return t.removeBucketsDirs(ctx, logger, shard, props, t.reindexBucketName)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) removeIngestBucketsDirs(ctx context.Context, logger logrus.FieldLogger,
-	shard ShardLike, props []string,
-) error {
-	return t.removeBucketsDirs(ctx, logger, shard, props, t.ingestBucketName)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) removeBucketsDirs(ctx context.Context, logger logrus.FieldLogger,
-	shard ShardLike, props []string, bucketNamer func(string) string,
-) error {
-	lsmPath := shard.pathLSM()
-	eg, _ := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
-	eg.SetLimit(t.config.concurrency)
-	for i := range props {
-		propName := props[i]
-
-		eg.Go(func() error {
-			bucketName := bucketNamer(propName)
-			bucketPath := filepath.Join(lsmPath, bucketName)
-
-			logger.WithField("bucket", bucketName).Debug("removing bucket")
-
-			return os.RemoveAll(bucketPath)
-		})
-	}
-	return eg.Wait()
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) duplicateToIngestBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string,
-) error {
-	return t.duplicateToBuckets(ctx, logger, shard, props, t.ingestBucketName, t.calcPropLenInverted)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) duplicateToMapBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string,
-) error {
-	return t.duplicateToBuckets(ctx, logger, shard, props, t.mapBucketName, t.calcPropLenMap)
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) duplicateToBuckets(ctx context.Context,
-	logger logrus.FieldLogger, shard ShardLike, props []string, bucketNamer func(string) string,
-	calcPropLen func([]inverted.Countable) float32,
-) error {
-	propsByName := map[string]struct{}{}
-	for i := range props {
-		propsByName[props[i]] = struct{}{}
-	}
-
-	shard.RegisterAddToPropertyValueIndex(func(s *Shard, docID uint64, property *inverted.Property) error {
-		if !property.HasSearchableIndex {
-			return nil
-		}
-		if _, ok := propsByName[property.Name]; !ok {
-			return nil
-		}
-
-		bucketName := bucketNamer(property.Name)
-		bucket := s.store.Bucket(bucketName)
-		propLen := calcPropLen(property.Items)
-		for _, item := range property.Items {
-			pair := s.pairPropertyWithFrequency(docID, item.TermFrequency, propLen)
-			if err := s.addToPropertyMapBucket(bucket, pair, item.Data); err != nil {
-				return fmt.Errorf("adding prop %q to bucket %q: %w", item.Data, bucketName, err)
-			}
-		}
-		return nil
-	})
-	shard.RegisterDeleteFromPropertyValueIndex(func(s *Shard, docID uint64, property *inverted.Property) error {
-		if !property.HasSearchableIndex {
-			return nil
-		}
-		if _, ok := propsByName[property.Name]; !ok {
-			return nil
-		}
-
-		bucketName := bucketNamer(property.Name)
-		bucket := s.store.Bucket(bucketName)
-		for _, item := range property.Items {
-			if err := s.deleteInvertedIndexItemWithFrequencyLSM(bucket, item, docID); err != nil {
-				return fmt.Errorf("deleting prop %q from bucket %q: %w", item.Data, bucketName, err)
-			}
-		}
-		return nil
-	})
-
-	return nil
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) isSegmentDb(filename string) bool {
-	return strings.HasPrefix(filename, "segment-") && strings.HasSuffix(filename, ".db")
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) isSegmentBloom(filename string) bool {
-	return strings.HasPrefix(filename, "segment-") && strings.HasSuffix(filename, ".bloom")
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) isSegmentWal(filename string) bool {
-	return strings.HasPrefix(filename, "segment-") && strings.HasSuffix(filename, ".wal")
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) calcPropLenMap(items []inverted.Countable) float32 {
-	return float32(len(items))
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) calcPropLenInverted(items []inverted.Countable) float32 {
-	propLen := float32(0)
-	for _, item := range items {
-		propLen += item.TermFrequency
-	}
-	return propLen
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) bucketOptions(shard ShardLike, strategy string,
-	disableCompaction, keepTombstones bool, memtableOptFactor int,
-) []lsmkv.BucketOption {
-	index := shard.Index()
-
-	opts := []lsmkv.BucketOption{
-		lsmkv.WithDirtyThreshold(time.Duration(index.Config.MemtablesFlushDirtyAfter) * time.Second),
-		lsmkv.WithDynamicMemtableSizing(
-			index.Config.MemtablesInitialSizeMB*memtableOptFactor,
-			index.Config.MemtablesMaxSizeMB*memtableOptFactor,
-			index.Config.MemtablesMinActiveSeconds*memtableOptFactor,
-			index.Config.MemtablesMaxActiveSeconds*memtableOptFactor,
-		),
-		lsmkv.WithPread(index.Config.AvoidMMap),
-		lsmkv.WithAllocChecker(index.allocChecker),
-		lsmkv.WithMaxSegmentSize(index.Config.MaxSegmentSize),
-		lsmkv.WithSegmentsChecksumValidationEnabled(index.Config.LSMEnableSegmentsChecksumValidation),
-		lsmkv.WithStrategy(strategy),
-		lsmkv.WithDisableCompaction(disableCompaction),
-		lsmkv.WithKeepTombstones(keepTombstones),
-	}
-
-	if strategy == lsmkv.StrategyMapCollection && shard.Versioner().Version() < 2 {
-		opts = append(opts, lsmkv.WithLegacyMapSorting())
-	}
-
-	return opts
-}
-
-func (t *ShardInvertedReindexTask_MapToBlockmax) ReindexByShard(ctx context.Context, shard ShardLike,
+func (t *ShardReindexTask_MapToBlockmax) OnAfterLsmInitAsync(ctx context.Context, shard ShardLike,
 ) (rerunAt time.Time, err error) {
 	collectionName := shard.Index().Config.ClassName.String()
 	logger := t.logger.WithFields(map[string]any{
 		"collection": collectionName,
 		"shard":      shard.Name(),
-		"method":     "ReindexByShard",
+		"method":     "OnAfterLsmInitAsync",
 	})
-	logger.Debug("starting")
+	logger.Info("starting")
 	defer func(started time.Time) {
 		logger = logger.WithField("took", time.Since(started))
 		if err != nil {
 			logger.WithError(err).Error("finished with error")
 		} else {
-			logger.Debug("finished")
+			logger.Info("finished")
 		}
 	}(time.Now())
 
 	zerotime := time.Time{}
 
 	if t.config.rollback {
+		logger.Debug("rollback. nothing to do")
 		return zerotime, nil
 	}
 
@@ -836,13 +340,13 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) ReindexByShard(ctx context.Cont
 		return zerotime, err
 	}
 
-	props, err := t.getPropsToReindex(shard, rt)
+	props, err := t.readPropsToReindex(rt)
 	if err != nil {
-		err = fmt.Errorf("getting reindexable props: %w", err)
+		err = fmt.Errorf("reading reindexable props: %w", err)
 		return zerotime, err
 	}
-	logger.WithField("props", props).Debug("props found")
 	if len(props) == 0 {
+		logger.Debug("no props read. nothing to do")
 		return zerotime, nil
 	}
 
@@ -967,19 +471,477 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) ReindexByShard(ctx context.Cont
 	return time.Now().Add(t.config.pauseDuration), nil
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) reindexBucketName(propName string) string {
+func (t *ShardReindexTask_MapToBlockmax) mergeReindexAndIngestBuckets(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, rt mapToBlockmaxReindexTracker, props []string,
+) error {
+	lsmPath := shard.pathLSM()
+	segmentPathsToMove := [][2]string{}
+	bucketPathsToRemove := make([]string, 0, len(props))
+	lock := new(sync.Mutex)
+
+	eg, gctx := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
+	eg.SetLimit(t.config.concurrency)
+	for i := range props {
+		propName := props[i]
+
+		eg.Go(func() error {
+			reindexBucketName := t.reindexBucketName(propName)
+			reindexBucketPath := filepath.Join(lsmPath, reindexBucketName)
+			ingestBucketName := t.ingestBucketName(propName)
+			ingestBucketPath := filepath.Join(lsmPath, ingestBucketName)
+
+			for {
+				propSegmentPathsToMove, needsRecover, err := t.getSegmentPathsToMove(reindexBucketPath, ingestBucketPath)
+				if err != nil {
+					return fmt.Errorf("buckets %q & %q: %w", reindexBucketName, ingestBucketName, err)
+				}
+
+				if needsRecover {
+					if err := t.recoverReindexBucket(gctx, logger, shard, reindexBucketName); err != nil {
+						return err
+					}
+				} else {
+					lock.Lock()
+					bucketPathsToRemove = append(bucketPathsToRemove, reindexBucketPath)
+					segmentPathsToMove = append(segmentPathsToMove, propSegmentPathsToMove...)
+					lock.Unlock()
+					return nil
+				}
+			}
+		})
+	}
+	if err := eg.Wait(); err != nil {
+		return err
+	}
+
+	logger.WithField("segments_to_move", segmentPathsToMove).WithField("buckets_to_remove", bucketPathsToRemove).
+		Debug("merging reindex and ingest buckets")
+
+	eg, _ = enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
+	eg.SetLimit(t.config.concurrency)
+	for i := range segmentPathsToMove {
+		i := i
+		eg.Go(func() error {
+			return os.Rename(segmentPathsToMove[i][0], segmentPathsToMove[i][1])
+		})
+	}
+	if err := eg.Wait(); err != nil {
+		return fmt.Errorf("moving segment: %w", err)
+	}
+
+	eg, _ = enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
+	eg.SetLimit(t.config.concurrency)
+	for i := range bucketPathsToRemove {
+		i := i
+		eg.Go(func() error {
+			return os.RemoveAll(bucketPathsToRemove[i])
+		})
+	}
+	if err := eg.Wait(); err != nil {
+		return fmt.Errorf("removing bucket: %w", err)
+	}
+
+	if err := rt.markMerged(); err != nil {
+		return fmt.Errorf("marking reindex merged: %w", err)
+	}
+	return nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) getSegmentPathsToMove(bucketPathSrc, bucketPathDst string,
+) ([][2]string, bool, error) {
+	segmentPaths := [][2]string{}
+	needsRecover := false
+
+	err := filepath.WalkDir(bucketPathSrc, func(path string, d os.DirEntry, err error) error {
+		if d.IsDir() {
+			return nil
+		}
+		if t.isSegmentDb(d.Name()) || t.isSegmentBloom(d.Name()) {
+			ext := filepath.Ext(d.Name())
+			id := strings.TrimSuffix(strings.TrimPrefix(d.Name(), "segment-"), ext)
+			timestamp, err := strconv.ParseInt(id, 10, 64)
+			if err != nil {
+				return err
+			}
+			timestampPast := time.Unix(0, timestamp).AddDate(-23, 0, 0).UnixNano()
+
+			segmentPaths = append(segmentPaths, [2]string{
+				path, filepath.Join(bucketPathDst, fmt.Sprintf("segment-%d%s", timestampPast, ext)),
+			})
+		} else if t.isSegmentWal(d.Name()) {
+			if info, err := d.Info(); err != nil {
+				return err
+			} else if info.Size() > 0 {
+				needsRecover = true
+				return filepath.SkipAll
+			}
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, false, err
+	}
+	if needsRecover {
+		return nil, true, nil
+	}
+	return segmentPaths, false, nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) swapIngestAndMapBuckets(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, rt mapToBlockmaxReindexTracker, props []string,
+) error {
+	lsmPath := shard.pathLSM()
+
+	eg, _ := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
+	eg.SetLimit(t.config.concurrency)
+	for i := range props {
+		propName := props[i]
+
+		if !rt.isSwappedProp(props[i]) {
+			eg.Go(func() error {
+				bucketName := helpers.BucketSearchableFromPropNameLSM(propName)
+				bucketPath := filepath.Join(lsmPath, bucketName)
+				ingestBucketName := t.ingestBucketName(propName)
+				ingestBucketPath := filepath.Join(lsmPath, ingestBucketName)
+				mapBucketName := t.mapBucketName(propName)
+				mapBucketPath := filepath.Join(lsmPath, mapBucketName)
+
+				logger.WithFields(map[string]any{
+					"bucket":        bucketName,
+					"ingest_bucket": ingestBucketName,
+					"map_bucket":    mapBucketName,
+				}).Debug("swapping buckets")
+
+				if err := os.Rename(bucketPath, mapBucketPath); err != nil {
+					return err
+				}
+				if err := os.Rename(ingestBucketPath, bucketPath); err != nil {
+					return err
+				}
+				if err := rt.markSwappedProp(propName); err != nil {
+					return fmt.Errorf("marking reindex swapped for %q: %w", propName, err)
+				}
+				return nil
+			})
+		}
+	}
+	if err := eg.Wait(); err != nil {
+		return err
+	}
+	if err := rt.markSwapped(); err != nil {
+		return fmt.Errorf("marking reindex swapped: %w", err)
+	}
+
+	logger.Debug("swapped searchable buckets")
+
+	return nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) unswapIngestAndMapBuckets(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, rt mapToBlockmaxReindexTracker, props []string,
+) error {
+	lsmPath := shard.pathLSM()
+
+	eg, _ := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
+	eg.SetLimit(t.config.concurrency)
+	for i := range props {
+		propName := props[i]
+
+		if rt.isSwappedProp(props[i]) {
+			eg.Go(func() error {
+				bucketName := helpers.BucketSearchableFromPropNameLSM(propName)
+				bucketPath := filepath.Join(lsmPath, bucketName)
+				ingestBucketName := t.ingestBucketName(propName)
+				ingestBucketPath := filepath.Join(lsmPath, ingestBucketName)
+				mapBucketName := t.mapBucketName(propName)
+				mapBucketPath := filepath.Join(lsmPath, mapBucketName)
+
+				logger.WithFields(map[string]any{
+					"bucket":        bucketName,
+					"ingest_bucket": ingestBucketName,
+					"map_bucket":    mapBucketName,
+				}).Debug("unswapping buckets")
+
+				if err := os.Rename(bucketPath, ingestBucketPath); err != nil {
+					return err
+				}
+				if err := os.Rename(mapBucketPath, bucketPath); err != nil {
+					return err
+				}
+				if err := rt.unmarkSwappedProp(propName); err != nil {
+					return fmt.Errorf("unmarking reindex swapped for %q: %w", propName, err)
+				}
+				return nil
+			})
+		}
+	}
+	if err := eg.Wait(); err != nil {
+		return err
+	}
+	if err := rt.unmarkSwapped(); err != nil {
+		return fmt.Errorf("unmarking reindex swapped: %w", err)
+	}
+
+	logger.Debug("unswapped searchable buckets")
+
+	return nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) tidyMapBuckets(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, rt mapToBlockmaxReindexTracker, props []string,
+) error {
+	lsmPath := shard.pathLSM()
+
+	eg, _ := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
+	eg.SetLimit(t.config.concurrency)
+	for i := range props {
+		propName := props[i]
+
+		eg.Go(func() error {
+			bucketName := t.mapBucketName(propName)
+			bucketPath := filepath.Join(lsmPath, bucketName)
+			if err := os.RemoveAll(bucketPath); err != nil {
+				return err
+			}
+			return nil
+		})
+	}
+
+	if err := eg.Wait(); err != nil {
+		return err
+	}
+	if err := rt.markTidied(); err != nil {
+		return fmt.Errorf("marking reindex tidied: %w", err)
+	}
+	return nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) loadReindexSearchBuckets(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, props []string,
+) error {
+	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyInverted, false, false, t.config.memtableOptBlockmaxFactor)
+	return t.loadBuckets(ctx, logger, shard, props, t.reindexBucketName, bucketOpts)
+}
+
+func (t *ShardReindexTask_MapToBlockmax) loadIngestSearchBuckets(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, props []string,
+	disableCompaction, keepTombstones bool,
+) error {
+	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyInverted, disableCompaction, keepTombstones, t.config.memtableOptBlockmaxFactor)
+	return t.loadBuckets(ctx, logger, shard, props, t.ingestBucketName, bucketOpts)
+}
+
+func (t *ShardReindexTask_MapToBlockmax) loadMapSearchBuckets(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, props []string,
+) error {
+	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyMapCollection, false, false, 1)
+	return t.loadBuckets(ctx, logger, shard, props, t.mapBucketName, bucketOpts)
+}
+
+func (t *ShardReindexTask_MapToBlockmax) loadBuckets(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, props []string, bucketNamer func(string) string,
+	bucketOpts []lsmkv.BucketOption,
+) error {
+	store := shard.Store()
+
+	eg, gctx := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
+	eg.SetLimit(t.config.concurrency)
+	for i := range props {
+		propName := props[i]
+
+		eg.Go(func() error {
+			bucketName := bucketNamer(propName)
+			logger.WithField("bucket", bucketName).Debug("loading bucket")
+			if err := store.CreateOrLoadBucket(gctx, bucketName, bucketOpts...); err != nil {
+				return err
+			}
+			logger.WithField("bucket", bucketName).Debug("bucket loaded")
+			return nil
+		})
+	}
+
+	if err := eg.Wait(); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) recoverReindexBucket(ctx context.Context,
+	logger logrus.FieldLogger, shard ShardLike, bucketName string,
+) error {
+	store := shard.Store()
+	bucketOpts := t.bucketOptions(shard, lsmkv.StrategyInverted, true, false, t.config.memtableOptBlockmaxFactor)
+
+	logger.WithField("bucket", bucketName).Debug("recover wals, loading bucket")
+	if err := store.CreateOrLoadBucket(ctx, bucketName, bucketOpts...); err != nil {
+		return fmt.Errorf("bucket %q: %w", bucketName, err)
+	}
+	logger.WithField("bucket", bucketName).Debug("recover wals, shutting down bucket")
+	if err := store.ShutdownBucket(ctx, bucketName); err != nil {
+		return fmt.Errorf("bucket %q: %w", bucketName, err)
+	}
+	logger.WithField("bucket", bucketName).Debug("recover wals, shut down bucket")
+
+	return nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) removeReindexBucketsDirs(ctx context.Context, logger logrus.FieldLogger,
+	shard ShardLike, props []string,
+) error {
+	return t.removeBucketsDirs(ctx, logger, shard, props, t.reindexBucketName)
+}
+
+func (t *ShardReindexTask_MapToBlockmax) removeIngestBucketsDirs(ctx context.Context, logger logrus.FieldLogger,
+	shard ShardLike, props []string,
+) error {
+	return t.removeBucketsDirs(ctx, logger, shard, props, t.ingestBucketName)
+}
+
+func (t *ShardReindexTask_MapToBlockmax) removeBucketsDirs(ctx context.Context, logger logrus.FieldLogger,
+	shard ShardLike, props []string, bucketNamer func(string) string,
+) error {
+	lsmPath := shard.pathLSM()
+	eg, _ := enterrors.NewErrorGroupWithContextWrapper(logger, ctx)
+	eg.SetLimit(t.config.concurrency)
+	for i := range props {
+		propName := props[i]
+
+		eg.Go(func() error {
+			bucketName := bucketNamer(propName)
+			bucketPath := filepath.Join(lsmPath, bucketName)
+
+			logger.WithField("bucket", bucketName).Debug("removing bucket")
+
+			return os.RemoveAll(bucketPath)
+		})
+	}
+	return eg.Wait()
+}
+
+func (t *ShardReindexTask_MapToBlockmax) duplicateToIngestBuckets(shard *Shard, props []string,
+) error {
+	return t.duplicateToBuckets(shard, props, t.ingestBucketName, t.calcPropLenInverted)
+}
+
+func (t *ShardReindexTask_MapToBlockmax) duplicateToMapBuckets(shard *Shard, props []string,
+) error {
+	return t.duplicateToBuckets(shard, props, t.mapBucketName, t.calcPropLenMap)
+}
+
+func (t *ShardReindexTask_MapToBlockmax) duplicateToBuckets(shard *Shard, props []string,
+	bucketNamer func(string) string, calcPropLen func([]inverted.Countable) float32,
+) error {
+	propsByName := map[string]struct{}{}
+	for i := range props {
+		propsByName[props[i]] = struct{}{}
+	}
+
+	shard.registerAddToPropertyValueIndex(func(s *Shard, docID uint64, property *inverted.Property) error {
+		if !property.HasSearchableIndex {
+			return nil
+		}
+		if _, ok := propsByName[property.Name]; !ok {
+			return nil
+		}
+
+		bucketName := bucketNamer(property.Name)
+		bucket := s.store.Bucket(bucketName)
+		propLen := calcPropLen(property.Items)
+		for _, item := range property.Items {
+			pair := s.pairPropertyWithFrequency(docID, item.TermFrequency, propLen)
+			if err := s.addToPropertyMapBucket(bucket, pair, item.Data); err != nil {
+				return fmt.Errorf("adding prop %q to bucket %q: %w", item.Data, bucketName, err)
+			}
+		}
+		return nil
+	})
+	shard.registerDeleteFromPropertyValueIndex(func(s *Shard, docID uint64, property *inverted.Property) error {
+		if !property.HasSearchableIndex {
+			return nil
+		}
+		if _, ok := propsByName[property.Name]; !ok {
+			return nil
+		}
+
+		bucketName := bucketNamer(property.Name)
+		bucket := s.store.Bucket(bucketName)
+		for _, item := range property.Items {
+			if err := s.deleteInvertedIndexItemWithFrequencyLSM(bucket, item, docID); err != nil {
+				return fmt.Errorf("deleting prop %q from bucket %q: %w", item.Data, bucketName, err)
+			}
+		}
+		return nil
+	})
+
+	return nil
+}
+
+func (t *ShardReindexTask_MapToBlockmax) isSegmentDb(filename string) bool {
+	return strings.HasPrefix(filename, "segment-") && strings.HasSuffix(filename, ".db")
+}
+
+func (t *ShardReindexTask_MapToBlockmax) isSegmentBloom(filename string) bool {
+	return strings.HasPrefix(filename, "segment-") && strings.HasSuffix(filename, ".bloom")
+}
+
+func (t *ShardReindexTask_MapToBlockmax) isSegmentWal(filename string) bool {
+	return strings.HasPrefix(filename, "segment-") && strings.HasSuffix(filename, ".wal")
+}
+
+func (t *ShardReindexTask_MapToBlockmax) calcPropLenMap(items []inverted.Countable) float32 {
+	return float32(len(items))
+}
+
+func (t *ShardReindexTask_MapToBlockmax) calcPropLenInverted(items []inverted.Countable) float32 {
+	propLen := float32(0)
+	for _, item := range items {
+		propLen += item.TermFrequency
+	}
+	return propLen
+}
+
+func (t *ShardReindexTask_MapToBlockmax) bucketOptions(shard ShardLike, strategy string,
+	disableCompaction, keepTombstones bool, memtableOptFactor int,
+) []lsmkv.BucketOption {
+	index := shard.Index()
+
+	opts := []lsmkv.BucketOption{
+		lsmkv.WithDirtyThreshold(time.Duration(index.Config.MemtablesFlushDirtyAfter) * time.Second),
+		lsmkv.WithDynamicMemtableSizing(
+			index.Config.MemtablesInitialSizeMB*memtableOptFactor,
+			index.Config.MemtablesMaxSizeMB*memtableOptFactor,
+			index.Config.MemtablesMinActiveSeconds*memtableOptFactor,
+			index.Config.MemtablesMaxActiveSeconds*memtableOptFactor,
+		),
+		lsmkv.WithPread(index.Config.AvoidMMap),
+		lsmkv.WithAllocChecker(index.allocChecker),
+		lsmkv.WithMaxSegmentSize(index.Config.MaxSegmentSize),
+		lsmkv.WithSegmentsChecksumValidationEnabled(index.Config.LSMEnableSegmentsChecksumValidation),
+		lsmkv.WithStrategy(strategy),
+		lsmkv.WithDisableCompaction(disableCompaction),
+		lsmkv.WithKeepTombstones(keepTombstones),
+	}
+
+	if strategy == lsmkv.StrategyMapCollection && shard.Versioner().Version() < 2 {
+		opts = append(opts, lsmkv.WithLegacyMapSorting())
+	}
+
+	return opts
+}
+
+func (t *ShardReindexTask_MapToBlockmax) reindexBucketName(propName string) string {
 	return helpers.BucketSearchableFromPropNameLSM(propName) + "__blockmax_reindex"
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) ingestBucketName(propName string) string {
+func (t *ShardReindexTask_MapToBlockmax) ingestBucketName(propName string) string {
 	return helpers.BucketSearchableFromPropNameLSM(propName) + "__blockmax_ingest"
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) mapBucketName(propName string) string {
+func (t *ShardReindexTask_MapToBlockmax) mapBucketName(propName string) string {
 	return helpers.BucketSearchableFromPropNameLSM(propName) + "__blockmax_map"
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) findPropsToReindex(shard ShardLike) []string {
+func (t *ShardReindexTask_MapToBlockmax) findPropsToReindex(shard ShardLike) []string {
 	propNames := []string{}
 	for name, bucket := range shard.Store().GetBucketsByName() {
 		if bucket.Strategy() == lsmkv.StrategyMapCollection && bucket.DesiredStrategy() == lsmkv.StrategyInverted {
@@ -996,7 +958,7 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) findPropsToReindex(shard ShardL
 	return propNames
 }
 
-func (t *ShardInvertedReindexTask_MapToBlockmax) getPropsToReindex(shard ShardLike, rt mapToBlockmaxReindexTracker) ([]string, error) {
+func (t *ShardReindexTask_MapToBlockmax) getPropsToReindex(shard ShardLike, rt mapToBlockmaxReindexTracker) ([]string, error) {
 	if rt.hasProps() {
 		props, err := rt.getProps()
 		if err != nil {
@@ -1011,6 +973,19 @@ func (t *ShardInvertedReindexTask_MapToBlockmax) getPropsToReindex(shard ShardLi
 	return props, nil
 }
 
+func (T *ShardReindexTask_MapToBlockmax) readPropsToReindex(rt mapToBlockmaxReindexTracker) ([]string, error) {
+	if rt.hasProps() {
+		props, err := rt.getProps()
+		if err != nil {
+			return nil, err
+		}
+		return props, nil
+	}
+	return []string{}, nil
+}
+
+// -----------------------------------------------------------------------------
+
 type migrationData struct {
 	key   indexKey
 	docID uint64
@@ -1018,424 +993,430 @@ type migrationData struct {
 	err   error
 }
 
-// type objectsIteratorAsync func(logger logrus.FieldLogger, shard ShardLike, lastKey indexKey, keyParse func([]byte) indexKey, propExtraction *storobj.PropertyExtraction, reindexStarted time.Time, breakCh <-chan bool,
-// ) (time.Time, <-chan *migrationData)
+type objectsIteratorAsync func(logger logrus.FieldLogger, shard ShardLike, lastKey indexKey, keyParse func([]byte) indexKey, propExtraction *storobj.PropertyExtraction, reindexStarted time.Time, breakCh <-chan bool,
+) (time.Time, <-chan *migrationData)
 
-// func uuidObjectsIteratorAsync(logger logrus.FieldLogger, shard ShardLike, lastKey indexKey, keyParse func([]byte) indexKey,
-// 	propExtraction *storobj.PropertyExtraction, reindexStarted time.Time, breakCh <-chan bool,
-// ) (time.Time, <-chan *migrationData) {
-// 	startedCh := make(chan time.Time)
-// 	mdCh := make(chan *migrationData)
+func uuidObjectsIteratorAsync(logger logrus.FieldLogger, shard ShardLike, lastKey indexKey, keyParse func([]byte) indexKey,
+	propExtraction *storobj.PropertyExtraction, reindexStarted time.Time, breakCh <-chan bool,
+) (time.Time, <-chan *migrationData) {
+	startedCh := make(chan time.Time)
+	mdCh := make(chan *migrationData)
 
-// 	enterrors.GoWrapper(func() {
-// 		cursor := shard.Store().Bucket(helpers.ObjectsBucketLSM).CursorOnDisk()
-// 		defer cursor.Close()
+	enterrors.GoWrapper(func() {
+		cursor := shard.Store().Bucket(helpers.ObjectsBucketLSM).CursorOnDisk()
+		defer cursor.Close()
 
-// 		startedCh <- time.Now() // after cursor created (necessary locks acquired)
-// 		addProps := additional.Properties{}
+		startedCh <- time.Now() // after cursor created (necessary locks acquired)
+		addProps := additional.Properties{}
 
-// 		var k, v []byte
-// 		if lastKey == nil {
-// 			k, v = cursor.First()
-// 		} else {
-// 			key := lastKey.Bytes()
-// 			k, v = cursor.Seek(key)
-// 			if bytes.Equal(k, key) {
-// 				k, v = cursor.Next()
-// 			}
-// 		}
+		var k, v []byte
+		if lastKey == nil {
+			k, v = cursor.First()
+		} else {
+			key := lastKey.Bytes()
+			k, v = cursor.Seek(key)
+			if bytes.Equal(k, key) {
+				k, v = cursor.Next()
+			}
+		}
 
-// 		for ; k != nil; k, v = cursor.Next() {
-// 			ik := keyParse(k)
-// 			obj, err := storobj.FromBinaryOptional(v, addProps, propExtraction)
-// 			if err != nil {
-// 				mdCh <- &migrationData{err: fmt.Errorf("unmarshalling object %q: %w", ik.String(), err)}
-// 				break
-// 			}
+		for ; k != nil; k, v = cursor.Next() {
+			ik := keyParse(k)
+			obj, err := storobj.FromBinaryOptional(v, addProps, propExtraction)
+			if err != nil {
+				mdCh <- &migrationData{err: fmt.Errorf("unmarshalling object %q: %w", ik.String(), err)}
+				break
+			}
 
-// 			if obj.LastUpdateTimeUnix() < reindexStarted.UnixMilli() {
-// 				props, _, err := shard.AnalyzeObject(obj)
-// 				if err != nil {
-// 					mdCh <- &migrationData{err: fmt.Errorf("analyzing object %q: %w", ik.String(), err)}
-// 					break
-// 				}
+			if obj.LastUpdateTimeUnix() < reindexStarted.UnixMilli() {
+				props, _, err := shard.AnalyzeObject(obj)
+				if err != nil {
+					mdCh <- &migrationData{err: fmt.Errorf("analyzing object %q: %w", ik.String(), err)}
+					break
+				}
 
-// 				if <-breakCh {
-// 					break
-// 				}
-// 				mdCh <- &migrationData{key: ik.Clone(), props: props, docID: obj.DocID}
-// 			} else {
-// 				if <-breakCh {
-// 					break
-// 				}
-// 				mdCh <- &migrationData{key: ik.Clone()}
-// 			}
-// 		}
-// 		if k == nil {
-// 			<-breakCh
-// 			mdCh <- nil
-// 		}
-// 		close(mdCh)
-// 	}, logger)
+				if <-breakCh {
+					break
+				}
+				mdCh <- &migrationData{key: ik.Clone(), props: props, docID: obj.DocID}
+			} else {
+				if <-breakCh {
+					break
+				}
+				mdCh <- &migrationData{key: ik.Clone()}
+			}
+		}
+		if k == nil {
+			<-breakCh
+			mdCh <- nil
+		}
+		close(mdCh)
+	}, logger)
 
-// 	return <-startedCh, mdCh
-// }
+	return <-startedCh, mdCh
+}
 
-// type reindexTracker interface {
-// 	isStarted() bool
-// 	markStarted(time.Time) error
-// 	getStarted() (time.Time, error)
+// -----------------------------------------------------------------------------
 
-// 	markProgress(lastProcessedKey indexKey, processedCount, indexedCount int) error
-// 	getProgress() (indexKey, error)
+type mapToBlockmaxReindexTracker interface {
+	isStarted() bool
+	markStarted(time.Time) error
+	getStarted() (time.Time, error)
 
-// 	isReindexed() bool
-// 	markReindexed() error
+	markProgress(lastProcessedKey indexKey, processedCount, indexedCount int) error
+	getProgress() (indexKey, error)
 
-// 	isMerged() bool
-// 	markMerged() error
+	isReindexed() bool
+	markReindexed() error
 
-// 	isSwapped() bool
-// 	markSwapped() error
-// 	unmarkSwapped() error
-// 	isSwappedProp(propName string) bool
-// 	markSwappedProp(propName string) error
-// 	unmarkSwappedProp(propName string) error
+	isMerged() bool
+	markMerged() error
 
-// 	isTidied() bool
-// 	markTidied() error
+	isSwapped() bool
+	markSwapped() error
+	unmarkSwapped() error
+	isSwappedProp(propName string) bool
+	markSwappedProp(propName string) error
+	unmarkSwappedProp(propName string) error
 
-// 	hasProps() bool
-// 	getProps() ([]string, error)
-// 	saveProps([]string) error
+	isTidied() bool
+	markTidied() error
 
-// 	reset() error
-// }
+	hasProps() bool
+	getProps() ([]string, error)
+	saveProps([]string) error
 
-// func newFileReindexTracker(lsmPath string, keyParser indexKeyParser) *fileReindexTracker {
-// 	return &fileReindexTracker{
-// 		progressCheckpoint: 1,
-// 		keyParser:          keyParser,
-// 		config: fileReindexTrackerConfig{
-// 			filenameStarted:    "started.mig",
-// 			filenameProgress:   "progress.mig",
-// 			filenameReindexed:  "reindexed.mig",
-// 			filenameMerged:     "merged.mig",
-// 			filenameSwapped:    "swapped.mig",
-// 			filenameTidied:     "tidied.mig",
-// 			filenameProperties: "properties.mig",
-// 			migrationPath:      filepath.Join(lsmPath, ".migrations", "searchable_map_to_blockmax"),
-// 		},
-// 	}
-// }
+	reset() error
+}
 
-// type fileReindexTracker struct {
-// 	progressCheckpoint int
-// 	keyParser          indexKeyParser
-// 	config             fileReindexTrackerConfig
-// }
+func newFileMapToBlockmaxReindexTracker(lsmPath string, keyParser indexKeyParser) *fileMapToBlockmaxReindexTracker {
+	return &fileMapToBlockmaxReindexTracker{
+		progressCheckpoint: 1,
+		keyParser:          keyParser,
+		config: fileReindexTrackerConfig{
+			filenameStarted:    "started.mig",
+			filenameProgress:   "progress.mig",
+			filenameReindexed:  "reindexed.mig",
+			filenameMerged:     "merged.mig",
+			filenameSwapped:    "swapped.mig",
+			filenameTidied:     "tidied.mig",
+			filenameProperties: "properties.mig",
+			migrationPath:      filepath.Join(lsmPath, ".migrations", "searchable_map_to_blockmax"),
+		},
+	}
+}
 
-// type fileReindexTrackerConfig struct {
-// 	filenameStarted    string
-// 	filenameProgress   string
-// 	filenameReindexed  string
-// 	filenameMerged     string
-// 	filenameSwapped    string
-// 	filenameTidied     string
-// 	filenameProperties string
-// 	migrationPath      string
-// }
+type fileMapToBlockmaxReindexTracker struct {
+	progressCheckpoint int
+	keyParser          indexKeyParser
+	config             fileReindexTrackerConfig
+}
 
-// func (t *fileReindexTracker) init() error {
-// 	if err := os.MkdirAll(t.config.migrationPath, 0o777); err != nil {
-// 		return err
-// 	}
-// 	return nil
-// }
+type fileReindexTrackerConfig struct {
+	filenameStarted    string
+	filenameProgress   string
+	filenameReindexed  string
+	filenameMerged     string
+	filenameSwapped    string
+	filenameTidied     string
+	filenameProperties string
+	migrationPath      string
+}
 
-// func (t *fileReindexTracker) isStarted() bool {
-// 	return t.fileExists(t.config.filenameStarted)
-// }
+func (t *fileMapToBlockmaxReindexTracker) init() error {
+	if err := os.MkdirAll(t.config.migrationPath, 0o777); err != nil {
+		return err
+	}
+	return nil
+}
 
-// func (t *fileReindexTracker) markStarted(started time.Time) error {
-// 	return t.createFile(t.config.filenameStarted, []byte(t.encodeTime(started)))
-// }
+func (t *fileMapToBlockmaxReindexTracker) isStarted() bool {
+	return t.fileExists(t.config.filenameStarted)
+}
 
-// func (t *fileReindexTracker) getStarted() (time.Time, error) {
-// 	path := t.filepath(t.config.filenameStarted)
-// 	content, err := os.ReadFile(path)
-// 	if err != nil {
-// 		return time.Time{}, err
-// 	}
-// 	return t.decodeTime(string(content))
-// }
+func (t *fileMapToBlockmaxReindexTracker) markStarted(started time.Time) error {
+	return t.createFile(t.config.filenameStarted, []byte(t.encodeTime(started)))
+}
 
-// func (t *fileReindexTracker) findLastProgressFile() (string, error) {
-// 	prefix := t.config.filenameProgress + "."
-// 	expectedLen := len(prefix) + 9 // 9 digits
+func (t *fileMapToBlockmaxReindexTracker) getStarted() (time.Time, error) {
+	path := t.filepath(t.config.filenameStarted)
+	content, err := os.ReadFile(path)
+	if err != nil {
+		return time.Time{}, err
+	}
+	return t.decodeTime(string(content))
+}
 
-// 	lastProgressFilename := ""
-// 	err := filepath.WalkDir(t.config.migrationPath, func(path string, d os.DirEntry, err error) error {
-// 		// skip parent and children dirs
-// 		if path != t.config.migrationPath {
-// 			if d.IsDir() {
-// 				return filepath.SkipDir
-// 			}
-// 			if name := d.Name(); len(name) == expectedLen && strings.HasPrefix(name, prefix) {
-// 				lastProgressFilename = name
-// 			}
-// 		}
-// 		return nil
-// 	})
+func (t *fileMapToBlockmaxReindexTracker) findLastProgressFile() (string, error) {
+	prefix := t.config.filenameProgress + "."
+	expectedLen := len(prefix) + 9 // 9 digits
 
-// 	return lastProgressFilename, err
-// }
+	lastProgressFilename := ""
+	err := filepath.WalkDir(t.config.migrationPath, func(path string, d os.DirEntry, err error) error {
+		// skip parent and children dirs
+		if path != t.config.migrationPath {
+			if d.IsDir() {
+				return filepath.SkipDir
+			}
+			if name := d.Name(); len(name) == expectedLen && strings.HasPrefix(name, prefix) {
+				lastProgressFilename = name
+			}
+		}
+		return nil
+	})
 
-// func (t *fileReindexTracker) markProgress(lastProcessedKey indexKey, processedCount, indexedCount int) error {
-// 	filename := fmt.Sprintf("%s.%09d", t.config.filenameProgress, t.progressCheckpoint)
-// 	content := strings.Join([]string{
-// 		t.encodeTime(time.Now()),
-// 		lastProcessedKey.String(),
-// 		fmt.Sprintf("all %d", processedCount),
-// 		fmt.Sprintf("idx %d", indexedCount),
-// 	}, "\n")
+	return lastProgressFilename, err
+}
 
-// 	if err := t.createFile(filename, []byte(content)); err != nil {
-// 		return err
-// 	}
-// 	t.progressCheckpoint++
-// 	return nil
-// }
+func (t *fileMapToBlockmaxReindexTracker) markProgress(lastProcessedKey indexKey, processedCount, indexedCount int) error {
+	filename := fmt.Sprintf("%s.%09d", t.config.filenameProgress, t.progressCheckpoint)
+	content := strings.Join([]string{
+		t.encodeTime(time.Now()),
+		lastProcessedKey.String(),
+		fmt.Sprintf("all %d", processedCount),
+		fmt.Sprintf("idx %d", indexedCount),
+	}, "\n")
 
-// func (t *fileReindexTracker) getProgress() (indexKey, error) {
-// 	filename, err := t.findLastProgressFile()
-// 	if err != nil {
-// 		return nil, err
-// 	}
-// 	if filename == "" {
-// 		return t.keyParser.FromBytes(nil), nil
-// 	}
+	if err := t.createFile(filename, []byte(content)); err != nil {
+		return err
+	}
+	t.progressCheckpoint++
+	return nil
+}
 
-// 	checkpoint, err := strconv.Atoi(strings.TrimPrefix(filename, t.config.filenameProgress+"."))
-// 	if err != nil {
-// 		return nil, err
-// 	}
+func (t *fileMapToBlockmaxReindexTracker) getProgress() (indexKey, error) {
+	filename, err := t.findLastProgressFile()
+	if err != nil {
+		return nil, err
+	}
+	if filename == "" {
+		return t.keyParser.FromBytes(nil), nil
+	}
 
-// 	path := t.filepath(filename)
-// 	content, err := os.ReadFile(path)
-// 	if err != nil {
-// 		return nil, err
-// 	}
+	checkpoint, err := strconv.Atoi(strings.TrimPrefix(filename, t.config.filenameProgress+"."))
+	if err != nil {
+		return nil, err
+	}
 
-// 	split := strings.Split(string(content), "\n")
-// 	key, err := t.keyParser.FromString(split[1])
-// 	if err != nil {
-// 		return nil, err
-// 	}
+	path := t.filepath(filename)
+	content, err := os.ReadFile(path)
+	if err != nil {
+		return nil, err
+	}
 
-// 	t.progressCheckpoint = checkpoint + 1
-// 	return key, nil
-// }
+	split := strings.Split(string(content), "\n")
+	key, err := t.keyParser.FromString(split[1])
+	if err != nil {
+		return nil, err
+	}
 
-// func (t *fileReindexTracker) isReindexed() bool {
-// 	return t.fileExists(t.config.filenameReindexed)
-// }
+	t.progressCheckpoint = checkpoint + 1
+	return key, nil
+}
 
-// func (t *fileReindexTracker) markReindexed() error {
-// 	return t.createFile(t.config.filenameReindexed, []byte(t.encodeTimeNow()))
-// }
+func (t *fileMapToBlockmaxReindexTracker) isReindexed() bool {
+	return t.fileExists(t.config.filenameReindexed)
+}
 
-// func (t *fileReindexTracker) isMerged() bool {
-// 	return t.fileExists(t.config.filenameMerged)
-// }
+func (t *fileMapToBlockmaxReindexTracker) markReindexed() error {
+	return t.createFile(t.config.filenameReindexed, []byte(t.encodeTimeNow()))
+}
 
-// func (t *fileReindexTracker) markMerged() error {
-// 	return t.createFile(t.config.filenameMerged, []byte(t.encodeTimeNow()))
-// }
+func (t *fileMapToBlockmaxReindexTracker) isMerged() bool {
+	return t.fileExists(t.config.filenameMerged)
+}
 
-// func (t *fileReindexTracker) isSwappedProp(propName string) bool {
-// 	return t.fileExists(t.config.filenameSwapped + "." + propName)
-// }
+func (t *fileMapToBlockmaxReindexTracker) markMerged() error {
+	return t.createFile(t.config.filenameMerged, []byte(t.encodeTimeNow()))
+}
 
-// func (t *fileReindexTracker) markSwappedProp(propName string) error {
-// 	return t.createFile(t.config.filenameSwapped+"."+propName, []byte(t.encodeTimeNow()))
-// }
+func (t *fileMapToBlockmaxReindexTracker) isSwappedProp(propName string) bool {
+	return t.fileExists(t.config.filenameSwapped + "." + propName)
+}
 
-// func (t *fileReindexTracker) unmarkSwappedProp(propName string) error {
-// 	return t.removeFile(t.config.filenameSwapped + "." + propName)
-// }
+func (t *fileMapToBlockmaxReindexTracker) markSwappedProp(propName string) error {
+	return t.createFile(t.config.filenameSwapped+"."+propName, []byte(t.encodeTimeNow()))
+}
 
-// func (t *fileReindexTracker) isSwapped() bool {
-// 	return t.fileExists(t.config.filenameSwapped)
-// }
+func (t *fileMapToBlockmaxReindexTracker) unmarkSwappedProp(propName string) error {
+	return t.removeFile(t.config.filenameSwapped + "." + propName)
+}
 
-// func (t *fileReindexTracker) markSwapped() error {
-// 	return t.createFile(t.config.filenameSwapped, []byte(t.encodeTimeNow()))
-// }
+func (t *fileMapToBlockmaxReindexTracker) isSwapped() bool {
+	return t.fileExists(t.config.filenameSwapped)
+}
 
-// func (t *fileReindexTracker) unmarkSwapped() error {
-// 	return t.removeFile(t.config.filenameSwapped)
-// }
+func (t *fileMapToBlockmaxReindexTracker) markSwapped() error {
+	return t.createFile(t.config.filenameSwapped, []byte(t.encodeTimeNow()))
+}
 
-// func (t *fileReindexTracker) isTidied() bool {
-// 	return t.fileExists(t.config.filenameTidied)
-// }
+func (t *fileMapToBlockmaxReindexTracker) unmarkSwapped() error {
+	return t.removeFile(t.config.filenameSwapped)
+}
 
-// func (t *fileReindexTracker) markTidied() error {
-// 	return t.createFile(t.config.filenameTidied, []byte(t.encodeTimeNow()))
-// }
+func (t *fileMapToBlockmaxReindexTracker) isTidied() bool {
+	return t.fileExists(t.config.filenameTidied)
+}
 
-// func (t *fileReindexTracker) filepath(filename string) string {
-// 	return filepath.Join(t.config.migrationPath, filename)
-// }
+func (t *fileMapToBlockmaxReindexTracker) markTidied() error {
+	return t.createFile(t.config.filenameTidied, []byte(t.encodeTimeNow()))
+}
 
-// func (t *fileReindexTracker) fileExists(filename string) bool {
-// 	if _, err := os.Stat(t.filepath(filename)); err == nil {
-// 		return true
-// 	} else if errors.Is(err, os.ErrNotExist) {
-// 		return false
-// 	}
-// 	return false
-// }
+func (t *fileMapToBlockmaxReindexTracker) filepath(filename string) string {
+	return filepath.Join(t.config.migrationPath, filename)
+}
 
-// func (t *fileReindexTracker) createFile(filename string, content []byte) error {
-// 	path := t.filepath(filename)
-// 	file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0o777)
-// 	if err != nil {
-// 		return err
-// 	}
-// 	defer file.Close()
+func (t *fileMapToBlockmaxReindexTracker) fileExists(filename string) bool {
+	if _, err := os.Stat(t.filepath(filename)); err == nil {
+		return true
+	} else if errors.Is(err, os.ErrNotExist) {
+		return false
+	}
+	return false
+}
 
-// 	if len(content) > 0 {
-// 		_, err = file.Write(content)
-// 		return err
-// 	}
-// 	return nil
-// }
+func (t *fileMapToBlockmaxReindexTracker) createFile(filename string, content []byte) error {
+	path := t.filepath(filename)
+	file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0o777)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
 
-// func (t *fileReindexTracker) removeFile(filename string) error {
-// 	return os.Remove(t.filepath(filename))
-// }
+	if len(content) > 0 {
+		_, err = file.Write(content)
+		return err
+	}
+	return nil
+}
 
-// func (t *fileReindexTracker) encodeTimeNow() string {
-// 	return t.encodeTime(time.Now())
-// }
+func (t *fileMapToBlockmaxReindexTracker) removeFile(filename string) error {
+	return os.Remove(t.filepath(filename))
+}
 
-// func (t *fileReindexTracker) encodeTime(tm time.Time) string {
-// 	return tm.UTC().Format(time.RFC3339Nano)
-// }
+func (t *fileMapToBlockmaxReindexTracker) encodeTimeNow() string {
+	return t.encodeTime(time.Now())
+}
 
-// func (t *fileReindexTracker) decodeTime(tm string) (time.Time, error) {
-// 	return time.Parse(time.RFC3339Nano, tm)
-// }
+func (t *fileMapToBlockmaxReindexTracker) encodeTime(tm time.Time) string {
+	return tm.UTC().Format(time.RFC3339Nano)
+}
 
-// func (t *fileReindexTracker) hasProps() bool {
-// 	return t.fileExists(t.config.filenameProperties)
-// }
+func (t *fileMapToBlockmaxReindexTracker) decodeTime(tm string) (time.Time, error) {
+	return time.Parse(time.RFC3339Nano, tm)
+}
 
-// func (t *fileReindexTracker) saveProps(propNames []string) error {
-// 	props := []byte(strings.Join(propNames, ","))
-// 	return t.createFile(t.config.filenameProperties, props)
-// }
+func (t *fileMapToBlockmaxReindexTracker) hasProps() bool {
+	return t.fileExists(t.config.filenameProperties)
+}
 
-// func (t *fileReindexTracker) getProps() ([]string, error) {
-// 	content, err := os.ReadFile(t.filepath(t.config.filenameProperties))
-// 	if err != nil {
-// 		return nil, err
-// 	}
-// 	if len(content) == 0 {
-// 		return []string{}, nil
-// 	}
-// 	return strings.Split(string(content), ","), nil
-// }
+func (t *fileMapToBlockmaxReindexTracker) saveProps(propNames []string) error {
+	props := []byte(strings.Join(propNames, ","))
+	return t.createFile(t.config.filenameProperties, props)
+}
 
-// func (t *fileReindexTracker) reset() error {
-// 	return os.RemoveAll(t.config.migrationPath)
-// }
+func (t *fileMapToBlockmaxReindexTracker) getProps() ([]string, error) {
+	content, err := os.ReadFile(t.filepath(t.config.filenameProperties))
+	if err != nil {
+		return nil, err
+	}
+	if len(content) == 0 {
+		return []string{}, nil
+	}
+	return strings.Split(string(content), ","), nil
+}
 
-// type indexKey interface {
-// 	String() string
-// 	Bytes() []byte
-// 	Clone() indexKey
-// }
+func (t *fileMapToBlockmaxReindexTracker) reset() error {
+	return os.RemoveAll(t.config.migrationPath)
+}
 
-// type uuidBytes []byte
+// -----------------------------------------------------------------------------
 
-// func (b uuidBytes) String() string {
+type indexKey interface {
+	String() string
+	Bytes() []byte
+	Clone() indexKey
+}
+
+type uuidBytes []byte
+
+func (b uuidBytes) String() string {
+	if b == nil {
+		return "nil"
+	}
+	uid, err := uuid.FromBytes(b)
+	if err != nil {
+		return err.Error()
+	}
+	return uid.String()
+}
+
+func (b uuidBytes) Bytes() []byte {
+	return b
+}
+
+func (b uuidBytes) Clone() indexKey {
+	buf := make([]byte, len(b))
+	copy(buf, b)
+	return uuidBytes(buf)
+}
+
+// type uint64Bytes []byte
+
+// func (b uint64Bytes) String() string {
 // 	if b == nil {
 // 		return "nil"
 // 	}
-// 	uid, err := uuid.FromBytes(b)
-// 	if err != nil {
-// 		return err.Error()
-// 	}
-// 	return uid.String()
+// 	return fmt.Sprint(binary.LittleEndian.Uint64(b))
 // }
 
-// func (b uuidBytes) Bytes() []byte {
+// func (b uint64Bytes) Bytes() []byte {
 // 	return b
 // }
 
-// func (b uuidBytes) Clone() indexKey {
+// func (b uint64Bytes) Clone() indexKey {
 // 	buf := make([]byte, len(b))
 // 	copy(buf, b)
-// 	return uuidBytes(buf)
+// 	return uint64Bytes(buf)
 // }
 
-// // type uint64Bytes []byte
+// -----------------------------------------------------------------------------
 
-// // func (b uint64Bytes) String() string {
-// // 	if b == nil {
-// // 		return "nil"
-// // 	}
-// // 	return fmt.Sprint(binary.LittleEndian.Uint64(b))
-// // }
+type indexKeyParser interface {
+	FromString(key string) (indexKey, error)
+	FromBytes(key []byte) indexKey
+}
 
-// // func (b uint64Bytes) Bytes() []byte {
-// // 	return b
-// // }
+type uuidKeyParser struct{}
 
-// // func (b uint64Bytes) Clone() indexKey {
-// // 	buf := make([]byte, len(b))
-// // 	copy(buf, b)
-// // 	return uint64Bytes(buf)
-// // }
+func (p *uuidKeyParser) FromString(key string) (indexKey, error) {
+	uid, err := uuid.Parse(key)
+	if err != nil {
+		return nil, err
+	}
+	buf, err := uid.MarshalBinary()
+	if err != nil {
+		return nil, err
+	}
+	return uuidBytes(buf), nil
+}
 
-// type indexKeyParser interface {
-// 	FromString(key string) (indexKey, error)
-// 	FromBytes(key []byte) indexKey
-// }
+func (p *uuidKeyParser) FromBytes(key []byte) indexKey {
+	return uuidBytes(key)
+}
 
-// type uuidKeyParser struct{}
+// type uint64KeyParser struct{}
 
-// func (p *uuidKeyParser) FromString(key string) (indexKey, error) {
-// 	uid, err := uuid.Parse(key)
+// func (p *uint64KeyParser) FromString(key string) (indexKey, error) {
+// 	u, err := strconv.ParseUint(key, 10, 64)
 // 	if err != nil {
 // 		return nil, err
 // 	}
-// 	buf, err := uid.MarshalBinary()
-// 	if err != nil {
-// 		return nil, err
-// 	}
-// 	return uuidBytes(buf), nil
+// 	buf := make([]byte, 8)
+// 	binary.LittleEndian.PutUint64(buf, u)
+// 	return uint64Bytes(buf), nil
 // }
 
-// func (p *uuidKeyParser) FromBytes(key []byte) indexKey {
-// 	return uuidBytes(key)
+// func (p *uint64KeyParser) FromBytes(key []byte) indexKey {
+// 	return uint64Bytes(key)
 // }
-
-// // type uint64KeyParser struct{}
-
-// // func (p *uint64KeyParser) FromString(key string) (indexKey, error) {
-// // 	u, err := strconv.ParseUint(key, 10, 64)
-// // 	if err != nil {
-// // 		return nil, err
-// // 	}
-// // 	buf := make([]byte, 8)
-// // 	binary.LittleEndian.PutUint64(buf, u)
-// // 	return uint64Bytes(buf), nil
-// // }
-
-// // func (p *uint64KeyParser) FromBytes(key []byte) indexKey {
-// // 	return uint64Bytes(key)
-// // }
