+++ adapters/repos/db/migrator.go
@@ -803,16 +803,6 @@ func (m *Migrator) InvertedReindex(ctx context.Context, taskNamesWithArgs map[st
 	return errs.ToError()
 }
 
-func (m *Migrator) ReindexerV2(tasksNames []string, args map[string]any) (*ReindexerV2, error) {
-	reindexer := NewReindexerV2(m.logger, m.db.indices, &m.db.indexLock)
-	for _, name := range tasksNames {
-		if err := reindexer.RegisterTask(name, args[name]); err != nil {
-			return nil, err
-		}
-	}
-	return reindexer, nil
-}
-
 func (m *Migrator) doInvertedReindex(ctx context.Context, taskNamesWithArgs map[string]any) error {
 	tasks := map[string]ShardInvertedReindexTask{}
 	for name, args := range taskNamesWithArgs {
